Set the file selectors filters according to the selected mode
When the Compress mode is selected, the source file selector's filter is set to
"All Files" and the destination file selector's filter is set to "PRS Files".
When the Decompress mode is selected, the opposite is done.

There was a bug in FileSelector that prevented setting its custom properties
after BeginInit was called. This bug is fixed by setting this.init to null in
EndInit.